Separate signing from verification
