+++ src/main/java/org/arkecosystem/crypto/signature/Verifier.java
@@ -1,0 +1,10 @@
+package org.arkecosystem.crypto.signature;
+
+
+import org.bitcoinj.core.ECKey;
+
+public interface Verifier {
+
+    boolean verify(byte[] hash, ECKey key, byte[] signature);
+
+}
