+++ src/test/java/org/arkecosystem/crypto/signature/ECDSASignerTest.java
@@ -6,7 +6,7 @@ import org.bitcoinj.core.ECKey;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class ECDSASignerTest {
 
@@ -27,23 +27,5 @@ class ECDSASignerTest {
             signature,
             Schnorr.bytesToHex(signer.sign(Schnorr.hexStringToByteArray(message), privateKey))
         );
-    }
-
-    @Test
-    void verify() {
-        byte[] message = Schnorr.hexStringToByteArray("243F6A8885A308D313198A2E03707344A4093822299F31D0082EFA98EC4E6C89");
-        ECKey privateKey = PrivateKey.fromPassphrase("some passphrase");
-        String signature = "304402207856D22D9C1E146492117B61D83F0A8D2E046C5A75F471172689A0A0C26907C6022035A1368F38EC63BC9B7BEEC7BCE49BD2C1AE28179EE83EBE7FD9C17824E2514B";
-
-        assertTrue(signer.verify(message, privateKey, Schnorr.hexStringToByteArray(signature)));
-    }
-
-    @Test
-    void verifyInvalid() {
-        byte[] message = Schnorr.hexStringToByteArray("243F6A8885A308D313198A2E03707344A4093822299F31D0082EFA98EC4E6C89");
-        ECKey privateKey = PrivateKey.fromPassphrase("some passphrase");
-        String signature = "00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001";
-
-        assertFalse(signer.verify(message, privateKey, Schnorr.hexStringToByteArray(signature)));
     }
 }
